+++ NEWS.md
@@ -1,3 +1,6 @@
+# detectMutation - hetre 0.1.9006
+* mutations cross-validation 2
+
 # detectMutation - hetre 0.1.9005
 * mutations cross-validation